feat: move to stable package
